+++ samples/assistant/powershell/GetChatState/run.ps1
@@ -1,0 +1,11 @@
+using namespace System.Net
+
+param($Request, $TriggerMetadata, $State)
+
+Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
+    StatusCode = [HttpStatusCode]::OK
+    Body       = $State
+    Headers    = @{
+        "Content-Type" = "application/json"
+    }
+})