+++ src/WebJobs.Extensions.OpenAI.AzureAISearch/CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v0.2.0 - Unreleased
+## v0.3.0 - Unreleased
+
+### Changed
+
+- Added support for key authentication back (use of managed identity is always recommended)
+- Updated Microsoft.Azure.WebJobs.Extensions.OpenAI to 0.15.0
+
+## v0.2.0 - 2024/04/24
 
 ### Changed